requirements.txt
streamlit==1.24.0
Pillow==9.5.0
pytesseract==0.3.10
openai==0.27.8